So Chrome devtools refactored their inspector network resource update messages, and their handling of redirects in WebKit.
This breaks our network view on the Chromium nightlies. This patch is a Bandaid
fix for Network resource translation layer that puts in a runtime swap of the
InspectorResourceConverter depending on the type of resource message we see
come out.

We have a major TODO to remove the legacy translation once WebKit r52154 gets
pulled into the Dev Channel. We also want to examine using a single record
format for everything since we are running into inband/outofband issues with
the timeline record enum.

R: zundel/ http://gwt-code-reviews.appspot.com/125804/show