Adds support for viewing Stack Traces and their source mappings to the UI.
This patch adds a SourceViewer widget whose job it is to fetch a remote
resource, format it, and expose API for getting individual lines of source.

Note that the code paths will be totally unused until the relevant changes
are made to Chromium to export the required data.

R: zundel / http://gwt-code-reviews.appspot.com/128806/show