This is the first part of a series of changes to add summary reports for information within the current window selection.
This patch refactors the old HintletReport to show only hints within the current
window selection.

The previous implementation showed all hintlets ever encountered in the
monitoring session. This was not very useful.

This patch refactors DataDispatcher to allow for time based lookups of all the
EventRecords. It also tweaks the report dialog to make way for future additions,
namely the upcoming summary piecharts.

R: conroy, zundel
http://gwt-code-reviews.appspot.com/615801/show