The previous version of the dashboard demo attempted to merge together app startup latency measurements with some measured actions on the UI.  In practice, it doesn't work well to try to view the same data set in these different ways, as each sample is either a page load measurement, or an action latency measurement.
The simplest thing I could think of was to split the demo into two dashboards.
The demo assumes that all the data in the datastore pertains to just one
dashboard or the other.

A more sophisticated demo might be able to discriminate between different types
of data in the datastore and present multiple dashboards on the same page or in
tabs.

R:conroy
http://gwt-code-reviews.appspot.com/618801/show