The is the second patch for the new summary report dialog.
This adds a summary pie chart that shows a breakdown of the time spent within
the selected window.

I changed the traversal for gathering the events within the window to walk from
right to left due to the fact that the binary search always returns the
insertion point to the right of the node we want to begin with, and also because
we sometimes have the left edge of the window falling in between some event.

R: conroy, zundel
http://gwt-code-reviews.appspot.com/617802/show