Updates Speedtracer to manifest V2
Chrome extensions now disallow inline scripts tags. There is no way around it,
so I had to redo the way Speedtracer gets linked and packaged.

This does not fix the brokenness on chrome 22 yet. This just gets us to a place
where we can develop locally with an unpacked extension.

TBR:kellegous
http://gwt-code-reviews.appspot.com/1846803/